+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 **Note** This version is not yet released and is under active development.
 
 ### Added
+- RTX-accelerated tetrahedrons geometry
+- generate aligned or randomly rotated data markers if some vectors are missong
+- methods to read back camera eye/target, light position, color and r/u/v
+- get_param() to read back the rt parameters
 - this changelog, markdown description content type tag for PyPI
 - use [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
-- read back camera eye/target, light position, color and r/u/v
 
 ## 0.1.2.5 - 2019-04-16
 Two weeks and some steps from the initial release. Starting changelog w.r.t. this point.